rag_pipeline/requirements.txt
# Core LangChain packages
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1

# OpenAI
openai>=1.0.0

# Vector stores
faiss-cpu>=1.7.4
chromadb>=0.4.0

# Web scraping
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0

# Environment and configuration
python-dotenv>=1.0.0

# Text processing and tokenization
tiktoken>=0.5.0

# Data handling
numpy>=1.24.0
pandas>=2.0.0

# Utilities
tqdm>=4.65.0
coloredlogs>=15.0

# Optional: FastAPI for web deployment
fastapi>=0.100.0
uvicorn>=0.23.0

# Optional: Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Optional: Jupyter for experimentation
jupyter>=1.0.0
ipykernel>=6.25.0